backend/requirements/websocket.txt
# 🔌 WebSocket Service Requirements
# Dependencias específicas para el servicio WebSocket

# Core WebSocket
websockets==12.0
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Async & Events
asyncio-mqtt==0.13.0
aioredis==2.0.1

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing (optional for WebSocket)
pytest==7.4.3
pytest-asyncio==0.21.1 